# Integrate SequenceBatcher with new threading abstraction