… src/core/sequence_batch_scheduler.h