There is a big performance slump.

**Config:**
model: simple_savedmodel_sequence_object
instance count: 1
max_batch_size: 8
sequence_batching {
  max_sequence_idle_microseconds: 5000000
}
concurrent sequences: 4
request concurrencies: 5


**New SequenceBatcher:**
Concurrency: 5, throughput: 1830 infer/sec, latency 2701 usec
Avg request latency: 2543 usec (overhead 91 usec + **queue 2008 usec** + compute input 32 usec + compute infer 403 usec + compute output 9 usec)

**Current SequenceBatcher**
Concurrency: 5, throughput: 3619 infer/sec, latency 1348 usec
Avg request latency: 1171 usec (overhead 178 usec + **queue 512 usec** + compute input 37 usec + compute infer 431 usec + compute output 13 usec)

The nsight traces  show some serious missed notification. I will address the perf slump as a part of another ticket. Let me know if something looks seriously flawed.